feat: add footer with github link
