+++ fresh.gen.ts
@@ -6,7 +6,9 @@ import * as $0 from "./routes/_404.tsx";
 import * as $1 from "./routes/_app.tsx";
 import * as $2 from "./routes/index.tsx";
 import * as $3 from "./routes/projects.tsx";
-import * as $$0 from "./islands/Header.tsx";
+import * as $$0 from "./islands/Card.tsx";
+import * as $$1 from "./islands/Footer.tsx";
+import * as $$2 from "./islands/Header.tsx";
 
 const manifest = {
   routes: {
@@ -16,7 +18,9 @@ const manifest = {
     "./routes/projects.tsx": $3,
   },
   islands: {
-    "./islands/Header.tsx": $$0,
+    "./islands/Card.tsx": $$0,
+    "./islands/Footer.tsx": $$1,
+    "./islands/Header.tsx": $$2,
   },
   baseUrl: import.meta.url,
 };
