+++ islands/Footer.tsx
@@ -1,0 +1,26 @@
+type FooterProps = {
+  title: string;
+  content: string | JSX.Element;
+};
+
+export const Footer = () => {
+  return (
+    <div class="container mx-auto px-8 mb-8">
+      <hr class="mx-auto my-8 max-w-md"/>
+      <div class="text-center">
+        <span class="inline-block">
+          <a href="https://github.com/devict/help" target="_blank" rel="noopener noreferrer" class="text-md font-medium">
+            <span class="flex items-center gap-1">
+              Github
+              <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="currentColor" className="w-6 h-6">
+                <path fillRule="evenodd" d="M15.75 2.25H21a.75.75 0 01.75.75v5.25a.75.75 0 01-1.5 0V4.81L8.03 17.03a.75.75 0 01-1.06-1.06L19.19 3.75h-3.44a.75.75 0 010-1.5zm-10.5 4.5a1.5 1.5 0 00-1.5 1.5v10.5a1.5 1.5 0 001.5 1.5h10.5a1.5 1.5 0 001.5-1.5V10.5a.75.75 0 011.5 0v8.25a3 3 0 01-3 3H5.25a3 3 0 01-3-3V8.25a3 3 0 013-3h8.25a.75.75 0 010 1.5H5.25z" clipRule="evenodd" />
+              </svg>
+            </span>
+          </a>
+        </span>
+      </div>
+    </div>
+  );
+};
+
+export default Footer;
