+++ routes/_app.tsx
@@ -1,5 +1,6 @@
 import { AppProps } from "$fresh/server.ts";
 import Header from "../islands/Header.tsx";
+import Footer from "../islands/Footer.tsx";
 
 const nav = [
   { name: "Home", href: "/" },
@@ -44,6 +45,7 @@ export default function App({ Component, route }: AppProps) {
             <Component />
           </div>
         </div>
+        <Footer />
       </body>
     </html>
   );
